Added Handshake class Fixed NetworkBuffer include Fixed some data types TcpClient: Added SendBytes, SetOnDisconnectCallback, SetOnConnectCallback, SetOnMessageCallback methods and made callback std::function's private and fixed some method implementations Added AcceptConnection on TcpServer Fixed TcpServer AddToClientsList

+++ include/NetworkMessage.hpp
@@ -12,7 +12,7 @@ struct NetworkMessage
 {
 	NetworkMessage();
 	NetworkMessage(const NetworkBuffer &buffer);
-	NetworkMessage(uint16 sender, byte distribution_mode, uint16 destination_id, byte tag, uint16 subject, NetworkBuffer buffer);
+	NetworkMessage(uint16 sender, byte distribution_mode, uint16 destination_id, byte tag, byte subject, NetworkBuffer buffer);
 	~NetworkMessage();
 
 	static const NetworkBuffer &EncodeMessage(const NetworkMessage &message);
@@ -23,7 +23,7 @@ struct NetworkMessage
 	byte distribution_mode;
 	uint16 destination_id;
 	byte tag;
-	uint16 subject;
+	byte subject;
 	void *data;
 	
 	bool valid = false;
